Restore long-press snooze sheet on glucose view
