+++ Trio/Sources/Modules/Home/View/HomeRootView.swift
@@ -34,6 +34,7 @@ extension Home {
         @State var selectedTab: Int = 0
         @State var showPumpSelection: Bool = false
         @State var showCGMSelection: Bool = false
+        @State var showSnoozeSheet: Bool = false
         @State var notificationsDisabled = false
         @State var timeButtons: [TimePicker] = [
             TimePicker(active: false, hours: 4),
@@ -141,6 +142,11 @@ extension Home {
                     } else {
                         state.shouldDisplayCGMSetupSheet.toggle()
                     }
+                }
+                .onLongPressGesture {
+                    let impactHeavy = UIImpactFeedbackGenerator(style: .heavy)
+                    impactHeavy.impactOccurred()
+                    showSnoozeSheet = true
                 }
         }
 
@@ -985,6 +991,9 @@ extension Home {
             }
             .sheet(isPresented: $state.isLegendPresented) {
                 ChartLegendView(state: state)
+            }
+            .sheet(isPresented: $showSnoozeSheet) {
+                SnoozeAllSheet(resolver: resolver, isPresented: $showSnoozeSheet)
             }
             // PUMP RELATED
             .confirmationDialog("Pump Model", isPresented: $showPumpSelection) {
