+++ Trio/Sources/Views/SnoozeAllSheet.swift
@@ -1,0 +1,73 @@
+import SwiftUI
+import Swinject
+
+/// Shared "Snooze All" sheet. Used from Notifications settings and the home
+/// glucose long-press. Wraps `TrioAlertManager.applySnooze` directly — no
+/// router / module hop.
+struct SnoozeAllSheet: View {
+    let resolver: Resolver
+    @Binding var isPresented: Bool
+
+    @State private var snoozeUntilDate: Date = .distantPast
+
+    @Environment(\.colorScheme) var colorScheme
+    @Environment(AppState.self) var appState
+
+    var body: some View {
+        NavigationStack {
+            List {
+                if snoozeUntilDate > Date() {
+                    Section {
+                        HStack {
+                            Image(systemName: "moon.zzz.fill").foregroundStyle(.tint)
+                            Text(String(
+                                format: String(localized: "Snoozed until %@"),
+                                snoozeUntilDate.formatted(date: .omitted, time: .shortened)
+                            ))
+                            .font(.headline)
+                        }
+                    }.listRowBackground(Color.chart)
+                }
+                Section(footer: Text(
+                    "Pick a duration to mute every Trio alarm. Critical alerts (e.g. occlusion, urgent low) still pierce the snooze."
+                )) {
+                    ForEach(NotificationResponseAction.allCases, id: \.self) { action in
+                        Button {
+                            applySnooze(action.duration)
+                        } label: {
+                            HStack {
+                                Text(action.localizedTitle).foregroundStyle(.primary)
+                                Spacer()
+                                Image(systemName: "chevron.right")
+                                    .foregroundColor(.gray)
+                                    .font(.footnote)
+                            }
+                        }
+                    }
+                }.listRowBackground(Color.chart)
+            }
+            .scrollContentBackground(.hidden)
+            .background(appState.trioBackgroundColor(for: colorScheme))
+            .navigationTitle("Snooze Alerts")
+            .navigationBarTitleDisplayMode(.inline)
+            .toolbar {
+                ToolbarItem(placement: .topBarTrailing) {
+                    Button("Close") { isPresented = false }
+                }
+            }
+            .onAppear {
+                snoozeUntilDate = UserDefaults.standard
+                    .object(forKey: "UserNotificationsManager.snoozeUntilDate") as? Date ?? .distantPast
+            }
+        }
+    }
+
+    private func applySnooze(_ duration: TimeInterval) {
+        let trioAlertManager = resolver.resolve(TrioAlertManager.self)
+        Task { @MainActor in
+            await trioAlertManager?.applySnooze(for: duration)
+            snoozeUntilDate = Date().addingTimeInterval(duration)
+            isPresented = false
+        }
+    }
+}
